separate out hubmap specs
