+++ src/layouts/MainLayout.vue
@@ -18,7 +18,8 @@
             icon="home"
           />
           <q-route-tab no-caps to="/Examples" label="Examples" />
-          <q-route-tab no-caps to="/HuBMAP" label="HuBMAP Examples" />
+          <q-route-tab no-caps to="/HuBMAPExamples" label="HuBMAP Examples" />
+          <q-route-tab no-caps to="/HuBMAPDemo" label="HuBMAP Demo" />
           <q-route-tab no-caps to="/Editor" label="Editor" />
         </q-tabs>
         <q-space />
